added Track file class
Similar to db::file::Index, the new db::file::Track class reads and writes the data structure representing a track. This includes the track's id, metadata and a list of its different media formats.

Replaced the part of db-add.cpp that writes the track file with the above class for simplicity.