fixed db file Track hanging upon read
When reading a track file, the db::file::Track class' read() function would hang due to it entering an infinity loop when attempting to skip empty lines (is.peek() called without subsequent is.get()). An additional is.peek() at the end of the loop ensures that an EOF is caught properly.

Also made the exception text more descriptive.