fixed Track read logic for format field
The read() function of db/file/Track.hpp does not properly read the format field of each format line. The buffer is filled immediately with characters, even those that are still in ctmp and might be part of the double hyphen separator (' -- ').

This commit fixes this by moving characters into the format buffer only after they drop out of the ctmp ringbuffer.