added db-hash-object and db-cat-object tools
The db-hash-object.cpp tool allows for an object to be hashed and written to the database. The db-cat-object.cpp tool allows for an object to be retrieved from the database by hash. The objects are prefixed with a header describing its type.

These two functions read/write objects in a objects/<hash:0:2>/<hash:2> structure. This is identical to git. In general, the above tools correspond vaguely to git hash-object and git cat-file, respectively.