db tools write object preambles and use db struct
The database tools db-add, db-list and db-meta now use the db/file/Object class to
read and write the preamble to their objects, i.e. the header that specifies the
object's type.

Said tools now also use the db/Structure object that provides access to the database
structure. This streamlines the name resolution process and makes the code more
readable and easier to maintain.