db/file classes read/write preamble themselves
The classes db/file/Index and db/file/Track now read and write the object preamble
inside their read() and write() functions. This makes using the classes easier,
since the caller does not have to account for the preamble. An additional benefit is
the proper stream checking when reading and writing db/file classes: The
read_preamble() and write_preamble() function act as a stream sanity check and throw
exceptions if something smells fishy.