added db/Object.hpp containing read/write tools
Added the db/Object.hpp file which contains the ObjectWriter class. The class
generalizes writing db/file/Generic objects into a buffer, calculating the hash from
the buffer's contents, and eventually writing the buffer to disk.

Multiple convenience overloads of write() are provided that allow the object to be
written to a regular path (char*), a Boost filesystem path (fs::path) or a location
in the database given by either a Structure::Object (if the hash is already known)
or a Structure (if the hash is to be used as object name). Especially the last
option is convenient, as it saves you from a lot of boilerplate code.