added ObjectReader to db/Object.hpp
Added the ObjectReader class to db/Object.hpp. Its purpose is to read an object from
a file on disk, capturing the object's hash (simply for reference) and providing an
interface that allows the caller to read() with exceptions, or maybe_read() with a
graceful `false` return value. This class, like ObjectWriter, makes the code of
other parts of the system shorter and more readable.