added db/ObjectBuffer that reads/writes objects
The ObjectBuffer class is a template that inherits from one of the db/file/* classes
and exposes the same interface. In addition, it encapsulates a ObjectReader and
ObjectWriter object and provides additional functions that streamline the process of
reading an object and writing the modified version back to disk.

To the user of the class it resembles the original db/file/* class, but in addition
keeps track of its hash when read (hash_in) and its hash when written (hash_out).
This allows an object to be read, modified and written back to disk, with the
programmer ending up with an object that captures both the hash before and after the
change.