track list in index now uses id as key
The track list in an index object now not only lists track object hashes, but
also the track id. The list is read into a map, with the track id as the key to the track object hash. This makes future track handling easier, where the
primary key to get to a track shall be the id (which is permanent), rather than
the hash (which changes with every modification).