added db-blob.cpp
Tool provides functions to display and modify a track's blobs. Particularly
useful to change the format or original name of a blob, and the only way to
actually add or remove new blogs. A conversion tool might use this command
to list tracks which lack a certain format, then fetch the blob from the
database, converting it to the desired format and adding the resulting blob
back to the track.